// generated/slskd-api/models/users-browse-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * slskd
 * A modern client-server application for the Soulseek file sharing network
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Directory } from './directory';
import {
    DirectoryFromJSON,
    DirectoryFromJSONTyped,
    DirectoryToJSON,
    DirectoryToJSONTyped,
} from './directory';

/**
 * 
 * @export
 * @interface UsersBrowseResponse
 */
export interface UsersBrowseResponse {
    /**
     * 
     * @type {Array<Directory>}
     * @memberof UsersBrowseResponse
     */
    directories?: Array<Directory>;
    /**
     * 
     * @type {number}
     * @memberof UsersBrowseResponse
     */
    directory_count?: number;
    /**
     * 
     * @type {Array<Directory>}
     * @memberof UsersBrowseResponse
     */
    locked_directories?: Array<Directory>;
    /**
     * 
     * @type {number}
     * @memberof UsersBrowseResponse
     */
    locked_directory_count?: number;
}

/**
 * Check if a given object implements the UsersBrowseResponse interface.
 */
export function instanceOfUsersBrowseResponse(value: object): value is UsersBrowseResponse {
    return true;
}

export function UsersBrowseResponseFromJSON(json: any): UsersBrowseResponse {
    return UsersBrowseResponseFromJSONTyped(json, false);
}

export function UsersBrowseResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): UsersBrowseResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'directories': json['directories'] == null ? undefined : ((json['directories'] as Array<any>).map(DirectoryFromJSON)),
        'directory_count': json['directoryCount'] == null ? undefined : json['directoryCount'],
        'locked_directories': json['lockedDirectories'] == null ? undefined : ((json['lockedDirectories'] as Array<any>).map(DirectoryFromJSON)),
        'locked_directory_count': json['lockedDirectoryCount'] == null ? undefined : json['lockedDirectoryCount'],
    };
}

export function UsersBrowseResponseToJSON(json: any): UsersBrowseResponse {
    return UsersBrowseResponseToJSONTyped(json, false);
}

export function UsersBrowseResponseToJSONTyped(value?: UsersBrowseResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'directories': value['directories'] == null ? undefined : ((value['directories'] as Array<any>).map(DirectoryToJSON)),
        'directoryCount': value['directory_count'],
        'lockedDirectories': value['locked_directories'] == null ? undefined : ((value['locked_directories'] as Array<any>).map(DirectoryToJSON)),
        'lockedDirectoryCount': value['locked_directory_count'],
    };
}